+++ src/cached.ts
@@ -1,17 +1,16 @@
-import type { AnyFn } from '../types'
-
 /**
  * 缓存执行结果, 当结果输入相同时, 不再执行
  * @param fn 方法
  */
-export const cached = <T extends AnyFn>(fn: T) => {
-  const caches = {}
-  return function() {
-    const args = Array.prototype.join.call(arguments, ',')
-    if (args in caches) {
-      return caches[args]
+export const cached = <T>(fn: (...args: any[]) => T) => {
+  const caches: Record<string, T> = {}
+  return function(...args: any[]) {
+    const argsStr = args.join(',')
+    if (argsStr in caches) {
+      return caches[argsStr]
     }
-    return (caches[args] = fn.apply(this, arguments))
-  } as T
+    caches[argsStr] = fn.apply(this, args)
+    return caches[argsStr]
+  }
 }